+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - `clamp()` is now available on the base `BigNumber` class
 
+👌 **Improvements**
+
+- Improved `@throws` exception documentation
+
 ## [0.14.6](https://github.com/brick/math/releases/tag/0.14.6) - 2026-02-05
 
 🗑️ **Deprecations**